WIP end-host redemption support refactor
